Simplify ChangeListener types
